fix(camera): sync issue
It seems like updateCamera() is occasionally run with a with a camera transformation value (map.transform) that cause the models to appear under the terrain layer. I don't know why this is, possibly a timing issue within mapbox? Updating the camera at the end of a move seems to ensure that they reappear. Not perfect as the models will still jitter and clip strangely while moving the camera but hopefully they will no longer be left in a broken state when the camera stops moving.